Extricate Foundation dependencies from JSON/CommonCodable; removal of global Date/Data strategies
