+++ Tests/NewCodableTests/CodableRevolutionTests.swift
@@ -60,17 +60,9 @@ struct NewCodableTests {
     
     @Test func dataDecoding() throws {
         let json = Data("[0, 1, 2, 3]".utf8)
-        let decoder = NewJSONDecoder(options: .init(dataDecodingStrategy: .deferredToData))
+        let decoder = NewJSONDecoder()
         let data = try decoder.decode(Data.self, from: json)
         #expect(data == Data([0, 1, 2, 3]))
-    }
-    
-    @Test func dataDecodingBase64() throws {
-        let json = Data("\"aGVsbG93b3JsZA==\"".utf8)
-        let decoder = NewJSONDecoder(options: .init(dataDecodingStrategy: .base64))
-        let b64DecodedData = try decoder.decode(Data.self, from: json)
-        let string = String(data: b64DecodedData, encoding: .utf8)
-        #expect(string == "helloworld")
     }
      
     @Test func embeddedCodable() throws {
@@ -1455,273 +1447,6 @@ struct NewCodableTests {
         let decoded = try NewJSONDecoder().decode(Aggregate.self, from: data)
         let expected = Aggregate(count: 15, suffixedStrings: ["a-foo", "b-foo", "c-foo", "d-foo"])
         #expect(decoded == expected)
-    }
-    
-    // TODO: Repeat with CommonCodable conformance.
-    @Test func codableAdaptorDataEncodingDecodingOptions() throws {
-        struct Test: JSONCodable, CommonCodable, Equatable {
-            struct Inner: Codable, Equatable {
-                let innerData: Data
-            }
-            
-            let outerData: Data
-            let inner: Inner
-            
-            static func decode(from decoder: inout some JSONDecoderProtocol & ~Escapable) throws(CodingError.Decoding) -> Test {
-                try decoder.decodeStruct { structDecoder throws(CodingError.Decoding) in
-                    var outerData: Data?
-                    var inner: Inner?
-                    try structDecoder.decodeEachKeyAndValue { key, valueDecoder throws(CodingError.Decoding) in
-                        switch key {
-                        case "outerData": 
-                            outerData = try valueDecoder.decode(Data.self)
-                        case "inner": 
-                            inner = try valueDecoder.decode(Inner.self)
-                        default: 
-                            break // Skip unknown keys
-                        }
-                        return false
-                    }
-                    
-                    guard let outerData, let inner else { 
-                        throw CodingError.dataCorrupted(debugDescription: "Missing required fields")
-                    }
-                    return Test(outerData: outerData, inner: inner)
-                }
-            }
-            
-            func encode(to encoder: inout JSONDirectEncoder) throws(CodingError.Encoding) {
-                try encoder.encodeDictionary(elementCount: 2) { dictEncoder throws(CodingError.Encoding) in
-                    try dictEncoder.encode(key: "outerData", value: outerData)
-                    try dictEncoder.encode(key: "inner", value: inner)
-                }
-            }
-            
-            static func decode(from decoder: inout some CommonDecoder & ~Escapable) throws(CodingError.Decoding) -> Test {
-                try decoder.decodeStruct { structDecoder throws(CodingError.Decoding) in
-                    var outerData: Data?
-                    var inner: Inner?
-                    try structDecoder.decodeEachKeyAndValue { key, valueDecoder throws(CodingError.Decoding) in
-                        switch key {
-                        case "outerData":
-                            outerData = try valueDecoder.decode(Data.self)
-                        case "inner":
-                            // Workaround, because CommonDecoder currently doesn't support Decodable.
-                            let dict = try valueDecoder.decode([String:Data].self)
-                            if let innerData = dict["innerData"] {
-                                inner = Inner(innerData: innerData)
-                            }
-                        default:
-                            break // Skip unknown keys
-                        }
-                        return false
-                    }
-                    
-                    guard let outerData, let inner else {
-                        throw CodingError.dataCorrupted(debugDescription: "Missing required fields")
-                    }
-                    return Test(outerData: outerData, inner: inner)
-                }
-            }
-            
-            func encode(to encoder: inout some CommonEncoder & ~Copyable & ~Escapable) throws(CodingError.Encoding) {
-                try encoder.encodeDictionary(elementCount: 1) { dictEncoder throws(CodingError.Encoding) in
-                    try dictEncoder.encode(key: "outerData", value: outerData)
-                    try dictEncoder.encode(key: "inner", value: ["innerData":inner.innerData]) // Workaroud, because CommonEncoder currently doesn't support Encodable.
-                }
-            }
-        }
-        
-        let original = Test(
-            outerData: Data(base64Encoded: "aGVsbG93b3JsZA==")!,
-            inner: .init(
-                innerData: Data(base64Encoded: "Z29vZGJ5ZQ==")!))
-        
-        let jsonData = """
-            {"outerData":"aGVsbG93b3JsZA==","inner":{"innerData":"Z29vZGJ5ZQ=="}}
-            """.data(using: .utf8)!
-
-        let encoder = NewJSONEncoder(options: .init(dataEncodingStrategy: .base64))
-        
-        @_transparent
-        func asJSON<D: JSONTopLevelEncoder & ~Copyable>(_ jsonEncoder: borrowing D) throws {
-            let data = try jsonEncoder.encode(original)
-            #expect(data == jsonData)
-        }
-        try asJSON(encoder)
-        
-        @_transparent
-        func asCommon<D: CommonTopLevelEncoder & ~Copyable>(_ commonEncoder: borrowing D) throws {
-            let data = try commonEncoder.encode(original)
-            #expect(data == jsonData)
-        }
-        try asCommon(encoder)
-        
-        let decoder = NewJSONDecoder(options: .init(dataDecodingStrategy: .base64))
-        
-        @_transparent
-        func asJSON<D: JSONTopLevelDecoder & ~Copyable>(_ jsonDecoder: borrowing D) throws {
-            let result = try jsonDecoder.decode(Test.self, from: jsonData)
-            #expect(result == original)
-        }
-        try asJSON(decoder)
-        
-        @_transparent
-        func asCommon<D: CommonTopLevelDecoder & ~Copyable>(_ commonDecoder: borrowing D) throws {
-            let result = try commonDecoder.decode(Test.self, from: jsonData)
-            #expect(result == original)
-        }
-        try asCommon(decoder)
-    }
-    
-    @Test func codableAdaptorDateEncodingDecodingOptions() throws {
-        struct Test: JSONCodable, CommonCodable, Equatable {
-            struct Inner: Codable, Equatable {
-                let innerDate: Date
-            }
-            
-            let outerDate: Date
-            let inner: Inner
-            
-            static func decode(from decoder: inout some JSONDecoderProtocol & ~Escapable) throws(CodingError.Decoding) -> Test {
-                try decoder.decodeStruct { structDecoder throws(CodingError.Decoding) in
-                    var outerDate: Date?
-                    var inner: Inner?
-                    
-                    try structDecoder.decodeEachKeyAndValue { key, valueDecoder throws(CodingError.Decoding) in
-                        switch key {
-                        case "outerDate": 
-                            outerDate = try valueDecoder.decode(Date.self)
-                        case "inner": 
-                            inner = try valueDecoder.decode(Inner.self)
-                        default: 
-                            break // Skip unknown keys
-                        }
-                        return false
-                    }
-                    
-                    guard let outerDate, let inner else { 
-                        throw CodingError.dataCorrupted(debugDescription: "Missing required fields")
-                    }
-                    return Test(outerDate: outerDate, inner: inner)
-                }
-            }
-            
-            func encode(to encoder: inout JSONDirectEncoder) throws(CodingError.Encoding) {
-                try encoder.encodeDictionary(elementCount: 2) { dictEncoder throws(CodingError.Encoding) in
-                    try dictEncoder.encode(key: "outerDate", value: outerDate)
-                    try dictEncoder.encode(key: "inner", value: inner)
-                }
-            }
-            
-            static func decode(from decoder: inout some CommonDecoder & ~Escapable) throws(CodingError.Decoding) -> Test {
-                try decoder.decodeStruct { structDecoder throws(CodingError.Decoding) in
-                    var outerDate: Date?
-                    var inner: Inner?
-                    
-                    try structDecoder.decodeEachKeyAndValue { key, valueDecoder throws(CodingError.Decoding) in
-                        switch key {
-                        case "outerDate":
-                            outerDate = try valueDecoder.decode(Date.self)
-                        case "inner":
-                            // Workaround, because CommonDecoder currently doesn't support Decodable.
-                            let dict = try valueDecoder.decode([String:Date].self)
-                            if let innerDate = dict["innerDate"] {
-                                inner = Inner(innerDate: innerDate)
-                            }
-                        default:
-                            break // Skip unknown keys
-                        }
-                        return false
-                    }
-                    
-                    guard let outerDate, let inner else {
-                        throw CodingError.dataCorrupted(debugDescription: "Missing required fields")
-                    }
-                    return Test(outerDate: outerDate, inner: inner)
-                }
-            }
-            
-            func encode(to encoder: inout some CommonEncoder & ~Copyable & ~Escapable) throws(CodingError.Encoding) {
-                try encoder.encodeDictionary(elementCount: 2) { dictEncoder throws(CodingError.Encoding) in
-                    try dictEncoder.encode(key: "outerDate", value: outerDate)
-                    try dictEncoder.encode(key: "inner", value: ["innerDate":inner.innerDate]) // Workaroud, because CommonEncoder currently doesn't support Encodable.
-                }
-            }
-        }
-        
-        let original = Test(
-            outerDate: Date(timeIntervalSince1970: 1),
-            inner: .init(
-                innerDate: Date(timeIntervalSince1970: 1)))
-        
-        // TODO: Figure out how to get the CommonEncoder/Decoder portions of this test to pass.
-        
-        func test(withEncodingStrategy encodingStrategy: NewJSONEncoder.Options.DateEncodingStrategy, decodingStrategy: NewJSONDecoder.Options.DateDecodingStrategy, expectedOutput: String) throws {
-            let encoder = NewJSONEncoder(options: .init(dateEncodingStrategy: encodingStrategy))
-            let decoder = NewJSONDecoder(options: .init(dateDecodingStrategy: decodingStrategy))
-                        
-            @_transparent
-            func asJSON<D: JSONTopLevelEncoder & ~Copyable>(_ jsonEncoder: borrowing D) throws -> Data{
-                let data = try jsonEncoder.encode(original)
-                let asString = String(data: data, encoding: .utf8)!
-                #expect(asString == expectedOutput)
-                return data
-            }
-            let data = try asJSON(encoder)
-            
-            @_transparent
-            func asCommon<D: CommonTopLevelEncoder & ~Copyable>(_ commonEncoder: borrowing D) throws {
-                let data = try commonEncoder.encode(original)
-                let asString = String(data: data, encoding: .utf8)!
-                #expect(asString == expectedOutput)
-            }
-            try asCommon(encoder)
-                        
-            @_transparent
-            func asJSON<D: JSONTopLevelDecoder & ~Copyable>(_ jsonDecoder: borrowing D) throws {
-                let result = try jsonDecoder.decode(Test.self, from: data)
-                #expect(result == original)
-            }
-            try asJSON(decoder)
-            
-            @_transparent
-            func asCommon<D: CommonTopLevelDecoder & ~Copyable>(_ commonDecoder: borrowing D) throws {
-                let result = try commonDecoder.decode(Test.self, from: data)
-                #expect(result == original)
-            }
-            try asCommon(decoder)
-        }
-        
-        let deferredToDateOutput = """
-            {"outerDate":-978307199,"inner":{"innerDate":-978307199}}
-            """
-        
-        try test(withEncodingStrategy: .deferredToDate, decodingStrategy: .deferredToDate, expectedOutput: deferredToDateOutput)
-        
-        let secondsSince1970Output = """
-            {"outerDate":1,"inner":{"innerDate":1}}
-            """
-        
-        try test(withEncodingStrategy: .secondsSince1970, decodingStrategy: .secondsSince1970, expectedOutput: secondsSince1970Output)
-        
-        let millisecondsSince1970Output = """
-            {"outerDate":1000,"inner":{"innerDate":1000}}
-            """
-        
-        try test(withEncodingStrategy: .millisecondsSince1970, decodingStrategy: .millisecondsSince1970, expectedOutput: millisecondsSince1970Output)
-
-        let iso8601Output = """
-            {"outerDate":"1970-01-01T00:00:01Z","inner":{"innerDate":"1970-01-01T00:00:01Z"}}
-            """
-        
-        try test(withEncodingStrategy: .iso8601, decodingStrategy: .iso8601, expectedOutput: iso8601Output)
-
-        let formatOutput = """
-            {"outerDate":"1970-01-01T00:00:01Z","inner":{"innerDate":"1970-01-01T00:00:01Z"}}
-            """
-        
-        try test(withEncodingStrategy: .formatted(Date.ISO8601FormatStyle()), decodingStrategy: .formatted(Date.ISO8601FormatStyle()), expectedOutput: formatOutput)
     }
     
     @Test func jsonNonConformingFloat() throws {
